requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-countries==7.2.1
django-crispy-forms==1.12.0
django-heroku==0.3.1
fontawesome-free==5.15.3
gunicorn==20.1.0
heroku==0.1.4
idna==2.10
Pillow==8.2.0
psycopg2==2.9.1
python-dateutil==1.5
python-decouple==3.4
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
